Refactor sql federation class and package
