+++ kernel/sql-federation/core/src/main/java/org/apache/shardingsphere/sqlfederation/optimizer/planner/rule/TranslatableProjectFilterRule.java
@@ -15,25 +15,29 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sqlfederation.optimizer.metadata.translatable;
+package org.apache.shardingsphere.sqlfederation.optimizer.planner.rule;
 
 import org.apache.calcite.plan.RelOptRule;
 import org.apache.calcite.plan.RelOptRuleCall;
 import org.apache.calcite.rel.core.RelFactories;
 import org.apache.calcite.rel.logical.LogicalFilter;
+import org.apache.calcite.rel.logical.LogicalProject;
 import org.apache.calcite.rex.RexCall;
+import org.apache.calcite.rex.RexInputRef;
 import org.apache.calcite.rex.RexNode;
 import org.apache.calcite.tools.RelBuilderFactory;
+import org.apache.shardingsphere.sqlfederation.optimizer.operator.TranslatableTableScan;
 
 import java.util.Collections;
+import java.util.List;
 import java.util.regex.Pattern;
 
 /**
- * Planner rule for pushing filters into table scan.
+ * Planner rule for pushing projections and filters into table scan.
  */
-public class TranslatableFilterRule extends RelOptRule {
+public class TranslatableProjectFilterRule extends RelOptRule {
     
-    public static final TranslatableFilterRule INSTANCE = new TranslatableFilterRule(RelFactories.LOGICAL_BUILDER);
+    public static final TranslatableProjectFilterRule INSTANCE = new TranslatableProjectFilterRule(RelFactories.LOGICAL_BUILDER);
     
     private static final Pattern CONDITION_PATTERN = Pattern.compile("\\$[A-Za-z]");
     
@@ -41,13 +45,38 @@ public class TranslatableFilterRule extends RelOptRule {
     
     private static final Pattern CONDITION_COMPLEX_PATTERN = Pattern.compile("NEGATED POSIX REGEX CASE SENSITIVE");
     
-    public TranslatableFilterRule(final RelBuilderFactory relBuilderFactory) {
-        super(operand(LogicalFilter.class, operand(TranslatableTableScan.class, none())), relBuilderFactory, "TranslatableFilterRule");
+    public TranslatableProjectFilterRule(final RelBuilderFactory relBuilderFactory) {
+        super(operand(LogicalProject.class, operand(LogicalFilter.class, operand(TranslatableTableScan.class, none()))), relBuilderFactory, "TranslatableProjectFilterRule");
+    }
+    
+    @Override
+    public void onMatch(final RelOptRuleCall call) {
+        LogicalProject project = call.rel(0);
+        int[] fields = getProjectFields(project.getProjects());
+        if (0 == fields.length) {
+            return;
+        }
+        LogicalFilter filter = call.rel(1);
+        TranslatableTableScan scan = call.rel(2);
+        call.transformTo(new TranslatableTableScan(scan.getCluster(), scan.getTable(), scan.getTranslatableTable(), Collections.singletonList(filter.getCondition()), fields));
+    }
+    
+    private int[] getProjectFields(final List<RexNode> rexNodes) {
+        int[] result = new int[rexNodes.size()];
+        for (int index = 0; index < rexNodes.size(); index++) {
+            RexNode exp = rexNodes.get(index);
+            if (exp instanceof RexInputRef) {
+                result[index] = ((RexInputRef) exp).getIndex();
+            } else {
+                return new int[0];
+            }
+        }
+        return result;
     }
     
     @Override
     public boolean matches(final RelOptRuleCall call) {
-        LogicalFilter filter = call.rel(0);
+        LogicalFilter filter = call.rel(1);
         RexCall condition = (RexCall) filter.getCondition();
         for (RexNode each : condition.getOperands()) {
             if (CONDITION_PATTERN.matcher(each.toString()).find()
@@ -57,12 +86,5 @@ public class TranslatableFilterRule extends RelOptRule {
             }
         }
         return true;
-    }
-    
-    @Override
-    public void onMatch(final RelOptRuleCall call) {
-        LogicalFilter filter = call.rel(0);
-        TranslatableTableScan scan = call.rel(1);
-        call.transformTo(new TranslatableTableScan(scan.getCluster(), scan.getTable(), scan.getTranslatableTable(), Collections.singletonList(filter.getCondition()), scan.getFields()));
     }
 }
